simplify package
