+++ requirements.txt
@@ -1,5 +1,3 @@
 ovos-plugin-manager>=0.8.5
 bm25s
 quebra-frases
-rapidfuzz
-json_database
